changes to run on python3

+++ TriggerAnalysis/plugins/L1MenuAnalyzer.cc
@@ -119,7 +119,7 @@ void L1MenuAnalyzer::analyze(const edm::Event& iEvent, const edm::EventSetup& iS
     const std::vector<std::pair<std::string, bool>> initialDecisions = gtUtil_->decisionsInitial();
     const std::vector<std::pair<std::string, bool>> intermDecisions = gtUtil_->decisionsInterm();
     const std::vector<std::pair<std::string, bool>> finalDecisions = gtUtil_->decisionsFinal();
-    const std::vector<std::pair<std::string, int>> prescales = gtUtil_->prescales();
+    const std::vector<std::pair<std::string, double>> prescales = gtUtil_->prescales();
     const std::vector<std::pair<std::string, std::vector<int>>> masks = gtUtil_->masks();
 
     if ((decisionCount_.size() != gtUtil_->decisionsInitial().size()) ||
